realtime-orders-system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
aiomysql==0.2.0
python-dotenv==1.0.0
pydantic==2.5.0
asyncio-mqtt==0.13.0